Major Refactor and Docs (#74)
- Miner Braiins fix div by zero, change layout, and add status
- Added guidance for creating nodeyez user, bitcoin permissions and lnd macaroon
- Refactored default image output to /home/nodeyez/nodeyez/imageoutput
- Refactored sample-config files from bitcoin to nodeyez
- Refactored service scripts from bitcoin to nodeyez
- Refactored nginx config from bitcoin to nodeyez
- Refactored script configFile and outputFile from bitcoin to nodeyez location
- Added individual doc files for each script, with configuration table
- Simplified README to be shorter with links to subpages for install
- Improved slideshow to restart based on image count when started to avoid alphabetical preference
- Set service wants for bitcoin and lnd service for those that depend on them

+++ scripts/slideshow.sh
@@ -2,9 +2,17 @@
 
 # the framebuffer image viewer (fbi) needs root so run this script with sudo
 # install fbi via  `sudo apt-get -y install fbi`
+foldertodisplay="/home/bitcoin/images"
+timeperimage=5
 while true
 do
-    fbi --vt 1 --autozoom --timeout 5 --device /dev/fb0 --noreadahead --cachemem 0 --noverbose --norandom /home/bitcoin/images/* > /dev/null 2>&1
-    sleep 120
-    killall fbi > /dev/null 2>&1
+    imagecount=`ls ${foldertodisplay} | wc -l`
+    sleepytime=$(($imagecount * $timeperimage))
+    if [ ${sleepytime} -gt 0 ]; then
+        fbi --vt 1 --autozoom --timeout ${timeperimage} --device /dev/fb0 --noreadahead --cachemem 0 --noverbose --norandom ${foldertodisplay}/* > /dev/null 2>&1
+        sleep ${sleepytime}
+        killall fbi > /dev/null 2>&1
+    else
+        sleep 15
+    fi
 done
